requirement.txt
uvicorn
fastapi
slackeventsapi
requests
python-dotenv
fastapi-utilities
